[YTA-1430] Add resized Worldpay logo and positioning style.
A page in the Payments project needs a logo to be placed inline with some text.
The existing Worldpay logo image file was a different size, and in Jpeg format.
The new logo is resized and in PNG format.
The added style aligns the logo vertically, to match a line of copy.
Change pixel measurements to ems.