Update back link
This commit updates the 'back' link to the latest version on GOV.UK
It introduces a new class .link-back. Everything is scoped to that class.
So this update will not affect any service unless they bump to the latest version
of the AF and also update the markup with the .link-back class

In this commit I have:
• Added new scoped styles for 'back' link to AF in govuk elements typography,
which matches where it lives in govuk assets
• Added the icon-arrow-left.png referenced in the the new styles for ie